+++ src/battle/data/effectRegistry.js
@@ -1,6 +1,7 @@
 // eslint-disable-next-line no-unused-vars
 const { effectIdEnum } = require("../../enums/battleEnums"); // TODO: remove after testing
 const { logger } = require("../../log");
+// eslint-disable-next-line no-unused-vars
 const types = require("../../../types");
 
 const allEffects = {};
@@ -48,7 +49,7 @@ const getEffect = (effectId) =>
   allEffects[effectId];
 
 /**
- * @param {Object} param0
+ * @param {object} param0
  * @param {Record<string, any>=} param0.fieldFilter
  * @param {Function=} param0.customFilter
  * @returns {types.PartialRecord<EffectIdEnum, Effect<any, any>>}
@@ -80,7 +81,7 @@ const getEffects = ({ fieldFilter, customFilter }) => {
   };
 };
 /**
- * @param {Object} param0
+ * @param {object} param0
  * @param {Record<string, any>=} param0.fieldFilter
  * @param {Function=} param0.customFilter
  * @returns {EffectIdEnum[]}
